Refactored to io.boomerang package
