+++ src/main/java/io/boomerang/AttributeConfig.java
@@ -1,13 +1,12 @@
-package net.boomerangplatform;
+package io.boomerang;
 
 import java.util.List;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import net.boomerangplatform.attributes.CloudEventAttributeArgumentResolver;
-import net.boomerangplatform.attributes.TokenAttributeArgumentResolver;
+import io.boomerang.attributes.CloudEventAttributeArgumentResolver;
+import io.boomerang.attributes.TokenAttributeArgumentResolver;
 
 @Configuration
 public class AttributeConfig implements WebMvcConfigurer {
