+++ src/main/java/io/boomerang/model/WebhookType.java
@@ -1,4 +1,4 @@
-package net.boomerangplatform.model;
+package io.boomerang.model;
 
 public enum WebhookType {
   generic, slack, dockerhub
