+++ src/main/java/io/boomerang/rest/config/RestConfig.java
@@ -1,4 +1,4 @@
-package net.boomerangplatform.rest.config;
+package io.boomerang.rest.config;
 
 import java.security.KeyManagementException;
 import java.security.KeyStoreException;
