+++ src/main/resources/application-local.properties
@@ -11,4 +11,4 @@ workflow.service.host=localhost:8080
 opentracing.jaeger.enabled=false
 
 # Logging
-logging.level.net.boomerangplatform=DEBUG
+logging.level.io.boomerang=DEBUG
